Bind DTO to container for using it like a FormRequest
