+++ src/Providers/ValidatedDTOServiceProvider.php
@@ -6,6 +6,7 @@ namespace WendellAdriel\ValidatedDTO\Providers;
 
 use Illuminate\Support\ServiceProvider;
 use WendellAdriel\ValidatedDTO\Console\Commands\MakeDTOCommand;
+use WendellAdriel\ValidatedDTO\Contracts\BaseDTO;
 
 final class ValidatedDTOServiceProvider extends ServiceProvider
 {
@@ -32,5 +33,16 @@ final class ValidatedDTOServiceProvider extends ServiceProvider
     public function register()
     {
         $this->mergeConfigFrom(__DIR__ . '/../../config/dto.php', 'dto');
+
+        $this->app->beforeResolving(BaseDTO::class, function ($class, $parameters, $app) {
+            if ($app->has($class)) {
+                return;
+            }
+
+            $app->bind(
+                $class,
+                fn ($container) => $class::fromRequest($container['request'])
+            );
+        });
     }
 }
